feat: create genres service
